requirements.txt
streamlit
langchain-groq
langchain
langchain-community
flask
python-dotenv
pypdf
langchain-groq
langchain_groq
langchain-huggingface